Update anaconda to 20.25.16
Apply result of "git diff anaconda-18.37.11-1..anaconda-20.25.16-1" and
resolve conflicts.

+++ anaconda/tests/pylint/runpylint.sh
@@ -4,58 +4,53 @@
 # of options minimizing false positives, in combination with filtering of any
 # warning regularexpressions listed in pylint-false-positives.
 # 
-# If any warnings our found they will be stored in pylint-log and printed
+# If any warnings are found they will be stored in pylint-log and printed
 # to stdout and this script will exit with a status of 1, if no (non filtered)
 # warnings are found it exits with a status of 0
 
-FALSE_POSITIVES=tests/pylint/pylint-false-positives
+# If $top_srcdir is set, assume this is being run from automake and we don't
+# need to keep a separate log
+pylint_log=0
+if [ -z "$top_srcdir" ]; then
+    pylint_log=1
+fi
 
-# W0107 - Unnecessary pass statement
-# W0108 - Lambda may not be necessary
+# Unset TERM so that things that use readline don't output terminal garbage
+unset TERM
+
+# If $top_srcdir has not been set by automake, import the test environment
+if [ -z "$top_srcdir" ]; then
+    top_srcdir="$(dirname "$0")/../.."
+    . ${top_srcdir}/tests/testenv.sh
+fi
+
+srcdir="${top_srcdir}/tests/pylint"
+
+FALSE_POSITIVES="${srcdir}"/pylint-false-positives
+
 # W0212 - Access to a protected member %s of a client class
-# W0311 - Bad indentation. Found %s %s, expected %s
-# W0312 - Found indentation with %ss instead of %ss
-# W0402 - Uses of a deprecated module %r
-# W0611 - Unused import %s
-# W0612 - Unused variable %r
-# W0710 - Exception doesn't inherit from standard "Exception" class
-NON_STRICT_OPTIONS="--disable=W0107,W0108,W0212,W0311,W0312,W0402,W0611,W0612,W0710"
+NON_STRICT_OPTIONS="--disable=W0212"
 
 # E1103 - %s %r has no %r member (but some types could not be inferred)
 DISABLED_ERR_OPTIONS="--disable=E1103"
 
-# W0102 - Dangerous default value %s as argument
+# W0110 - map/filter on lambda could be replaced by comprehension
 # W0141 - Used builtin function %r
 # W0142 - Used * or ** magic
-# W0201 - Attribute %r defined outside __init__
-# W0221 - Arguments number differs from %s method
 # W0223 - Method %r is abstract in class %r but is not overridden
-# W0231 - __init__ method from base class %r is not called
-# W0232 - Class has no __init__ method
-# W0233 - __init__ method from a non direct base class %r is called
-# W0401 - Wildcard import %s
-# W0403 - Relative import %r, should be %r
-# W0404 - Reimport %r (imported line %s)
 # W0511 - Used when a warning note as FIXME or XXX is detected.
-# W0602 - Using global for %r but no assignment is done
 # W0603 - Using the global statement
 # W0604 - Using the global statement at the module level
 # W0613 - Unused argument %r
 # W0614 - Unused import %s from wildcard import
-# W0621 - Redefining name %r from outer scope (line %s)
-# W0622 - Redefining built-in %r
-# W0702 - No exception type(s) specified
-# W0703 - Catch "Exception"
-# W1001 - Use of "property" on an old style class
-DISABLED_WARN_OPTIONS="--disable=W0102,W0141,W0142,W0201,W0221,W0223,W0231,W0232 \
-                       --disable=W0233,W0401,W0403,W0404,W0511,W0602,W0603,W0604 \
-                       --disable=W0613,W0614,W0621,W0622,W0702,W0703,W1001"
+DISABLED_WARN_OPTIONS="--disable=W0110,W0141,W0142,W0223,W0511,W0603,W0604,W0613,W0614"
 
 usage () {
-  echo "usage: `basename $0` [--strict] [--help]"
+  echo "usage: `basename $0` [--strict] [--help] [files...]"
   exit $1
 }
 
+FILES=
 while [ $# -gt 0 ]; do
   case $1 in
     --strict)
@@ -65,8 +60,8 @@ while [ $# -gt 0 ]; do
       usage 0
       ;;
     *)
-      echo "Error unknown option: $1"
-      usage 1
+      FILES=$@
+      break
   esac
   shift
 done
@@ -77,36 +72,56 @@ if [ "`tail -c 1 $FALSE_POSITIVES`" == "`echo`" ]; then
   exit 1
 fi
 
+exit_status=0
+if [ "$pylint_log" -ne 0 ]; then
+    > pylint-log
+fi
+
 # run pylint one file / module at a time, otherwise it sometimes gets
 # confused
-> pylint-log
-for i in $(find pyanaconda -type f -name '*py'); do
-  pylint --init-hook='import sys; \
-      sys.path.insert(1, "pyanaconda/isys/.libs"); \
-      sys.path.insert(2, "pyanaconda/isys"); \
-      sys.path.insert(3, "pyanaconda"); \
-      sys.path.insert(4, "pyanaconda/.libs"); \
-      sys.path.insert(5, "pyanaconda/iw"); \
-      sys.path.insert(6, "pyanaconda/textw"); \
-      sys.path.insert(7, "/usr/share/system-config-date"); \
-      sys.path.insert(8, "/usr/share/system-config-keyboard")' \
-    -i y -r n --disable=C,R --rcfile=/dev/null \
+if [ -z "$FILES" ]; then
+    # Find any file in the list of directories that either ends in .py
+    # or contains #!/usr/bin/python in the first line.
+    FILES="$(find "${top_srcdir}"/{anaconda,pyanaconda,tests,widgets,utils,scripts} -type f \( -name '*.py' -o -exec awk -e 'NR==1 { if ($0 ~ /^#!\/usr\/bin\/python/) exit 0; else exit 1; }' -e 'END { if (NR == 0) exit 1; }' {} \; \) -print)"
+fi
+for i in $FILES; do
+  if [ -n "$(echo "$i" | grep 'pyanaconda/packaging/dnfpayload.py$')" ]; then
+     continue
+  fi
+
+  pylint_output="$(pylint \
+    --msg-template='{msg_id}:{line:3d},{column}: {obj}: {msg}' \
+    -r n --disable=C,R --rcfile=/dev/null \
+    --dummy-variables-rgx=_ \
+    --ignored-classes=DefaultInstall,Popen,QueueFactory,TransactionSet \
+    --defining-attr-methods=__init__,_grabObjects,initialize,reset,start,setUp \
     $DISABLED_WARN_OPTIONS \
     $DISABLED_ERR_OPTIONS \
     $NON_STRICT_OPTIONS $i | \
-    egrep -v "`cat $FALSE_POSITIVES | tr '\n' '|'`" > pylint-tmp-log
-  if grep -q -v '************* Module ' pylint-tmp-log; then
-    cat pylint-tmp-log >> pylint-log
+    egrep -v "$(tr '\n' '|' < "$FALSE_POSITIVES") \
+    ")"
+  # I0011 is the informational "Locally disabling ...." message
+  if [ -n "$(echo "$pylint_output" | fgrep -v '************* Module ' |\
+          grep -v '^I0011:')" ]; then
+      # Replace the Module line with the actual filename
+      pylint_output="$(echo "$pylint_output" | sed "s|\* Module .*|* Module $i|")"
+
+      if [ "$pylint_log" -ne 0 ]; then
+          echo "$pylint_output" >> pylint-log
+      else
+          echo "$pylint_output"
+      fi
+      exit_status=1
   fi
 done
-rm pylint-tmp-log
 
-if [ -s pylint-log ]; then
-  echo "pylint reports the following issues:"
-  cat pylint-log
-  exit 1
+if [ "$pylint_log" -ne 0 ]; then
+    if [ -s pylint-log ]; then
+        echo "pylint reports the following issues:"
+        cat pylint-log
+    else
+        rm pylint-log
+    fi
 fi
 
-rm pylint-log
-
-exit 0
+exit "$exit_status"
